pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gif-framework"
version = "1.0.0"
description = "Genomic Intelligence Framework — Multi-component risk assessment for Listeria monocytogenes"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.9"
authors = [
    {name = "Javier Gamboa", email = "jgamboa@biotecno.org"},
]
keywords = ["genomics", "listeria", "food-safety", "risk-assessment", "WGS"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
]
dependencies = [
    "click>=8.0",
    "biopython>=1.80",
    "pandas>=2.0",
    "numpy>=1.24",
    "rich>=13.0",
    "requests>=2.28",
]

[project.scripts]
gif = "gif.cli:main"

[project.urls]
Homepage = "https://github.com/jgamboa-biotecno/GIF-Framework"
Repository = "https://github.com/jgamboa-biotecno/GIF-Framework"
Issues = "https://github.com/jgamboa-biotecno/GIF-Framework/issues"
